requirements.txt
bitsandbytes
transformers
peft
accelerate
datasets
scipy
einops
evaluate
trl
rouge-score
wandb
google-cloud-bigquery
